move example to docs
